+++ lib/repository/repository.dart
@@ -55,6 +55,7 @@ extension RepositoryActions on Repository {
 extension RepositoryCommons on Repository {
   String get repositoryPath => join(path, repositoryWorkingDirName);
   Directory get repositoryDirectory => Directory(repositoryPath);
+  Directory get workingDirectory => Directory(repositoryDirectory.parent.path);
   bool get isInitialized => repositoryDirectory.existsSync();
 }