+++ CHANGELOG.md
@@ -13,4 +13,5 @@
 - View commit history
 - Git diff on branches
 - Merge another branches commit into current working dir
-- Command line parsing upgrade
+- Command line parsing upgrade using args package
+-